refactor(docker-tests): move qtum_docker_node to helper layer
Ensures no test module depends on raw Docker patterns by moving
qtum_docker_node() function and Docker image constants from
qrc20_tests.rs to helpers/qrc20.rs.

Changes:
- Moved qtum_docker_node() to helpers/qrc20.rs (Docker node setup section)
- Moved QTUM_REGTEST_DOCKER_IMAGE constants to helpers/qrc20.rs (top of file)
- Removed raw Docker imports from qrc20_tests.rs (Command, testcontainers)
- Updated imports in qrc20_tests.rs and docker_tests_main.rs
- Pattern matches existing helpers (utxo.rs, zcoin.rs, eth.rs)

Verification:
- No raw Docker patterns (Command::new("docker")) in any *_tests.rs files
- All Docker operations now encapsulated in helper modules
- Code compiles without errors or warnings

Updated docs/plans/docker-tests-split.md to mark task complete.
